Add streaming search
A very rough initial support for streaming search, where the
initial request returns the id used for retrieving subsequent
chunks.

To keep everything within the confines of tokio_proto, every
streaming request has an associated protocol::StreamExchange,
which keeps a FIFO of received chunks and the sender half of
the channel to an outstanding chunk receiver, if any. Every
received chunk is pushed to the FIFO, and a small body frame
is returned with the original request id to keep tokio_proto
happy.

The client uses streaming_chunk() to request the next chunk. If
there are frames in the FIFO, the first one is returned; if there
are none, a receiver future is returned, which will resolve to
the first subsequently received chunk.

The code doesn't clean up the exchange map yet. Full LDAP
response codes/messages are absent, and so are controls in
either direction. The only way to recognize the end of the
search is to check the returned attribute map for zero size.