WIP rewrite to use the ordinary multiplexed client
The previous approach broke down when the number of streaming
chunks exceeded the capacity of tokio-proto's frame buffer
(MAX_BUFFERED_FRAMES = 128). Now everything is handled by
multiplexed::ClientProto.

All searches are streamed; a search request creates a MPSC queue
which receives the frames; the first response to the request
creates the stream struct which polls this queue. All search
entries have to be retrieved by polling the queue. Adapters
which hide the streaming shouldn't be too difficult to provide.

This is _very_ rough and not really usable, aside from limited
testing which showed that a stream of 500 results could be
received without overflowing any internal buffer.