Handle binary-valued attributes in results
If an attribute is an unconstrained octet string, its conversion
to String will fail. We now place all such attributes into a
separate hashmap. This is not ideal, as it _can_ conceivably
happen that some sets of values are successfully converted into
Strings, so the attribute must be looked for in two places, but
it's better than crashing.