Add timeout support
Timeouts are supported both for connection establishment (except
with Unix domain sockets) and for LDAP operations.

A connection must be built with LdapConnBuilder in order to be
timeout-capable while being opened. A timeout during connection
establishment will return an error wrapped in io::Error to
the caller.

An LDAP operation can have a timeout imposed by calling
with_timeout() on the connection handle before invoking the
operation. The same interface is used for synchronous and
asynchronous connections. If a timeout occurs, the returned
LdapResult will have a special status code indicating that fact.
The error will _not_ be returned through the futures chain; this
makes the connection usable for further use.

For Search operations, the timeout is reset for each expected
result entry. If a timeout occurs, the search is immediately
abandoned and a LdapResult with the timeout status code is
returned, along with all entries which have been collected so
far.