Redo timeout handling and resource management
Timeouts disrupt the regular resource management cycle inside
tokio-proto, which our library depends on. The chief problem is
that tokio-proto will not remove a protocol exchange from an
internal hashmap unless it receives a response with an appropriate
id; if a timeout occurs, that response may or may not arrive over
the network later, but it must be ignored by the protocol machinery.
Therefore, we must return a fake response in order to avoid a
small memory leak for each timeout.

Also, since we must have a separate mapping between LDAP message
ids (msgid) and tokio-proto ids (id) in order to conform to the
LDAP standard, entries in the (msgid->id) hashmap must also be freed
when a response arrives or a timeout occurs. (This is complicated
by the fact that most search responses arrive in multiple LDAP
messages sharing the same msgid, and are handled by the library
outside of regular tokio-proto exchanges.)

Furthemore, LDAP Abandon and Unbind requests never expect a
response, and can be handled in some way anologously to a
timed-out request wrt resource management.

This patch reworks and consolidates resource handling when
confronted with timeouts. The library should now be resource
neutral in all but a few edge cases involving TCP sending buffer
overflows (unconfirmed because it's difficult to test for.)