WIP Abandon rework
The Abandon operation can now be called directly on the synchronous
search stream or the channel obtained through an asynchronous
stream struct. If the Abandon message is succesfully sent to the
server, the next polling of the search stream will return an
error.

This is a departure from the way in which the operation result has
been communicated to the initiator (LdapResult with a nonstandard
result code). If it proves to be too unergonomic, it can be reverted.

Ldap::abandon() now accepts the LDAP message id as the argument.
Since the rest of the library provides no way to obtain the msgid,
it effectively means that the function can only be called
internally. It remains publicly visible, however, to reduce
churn if a later refactor makes it more universally usable.