The most significant bit in a INTEGER is always 0
When the number is positive.

If a INTEGER is only decoded with 8 bits one octet, and the
msb of that octet is 1, AD reads this as a signed integer,
and will therefore read a negative number.

This fixes that by incrementing the number of octets in a integer
if the msb is 1 and the number is positive.

Fixes #21