Prohibit parallel use of EntryStream and LdapConn
Those two structs share the Tokio runtime through an Arc. For each
operation a &mut of the runtime is taken from the Arc, which only
works if there is no other outstanding borrow. The runtime isn't
behind a Mutex since it would have be locked across an await point,
which should be avoided for the Mutex from std.

Therefore, we virtually capture a reference to LdapConn when making
an EntryStream, so spawning a free thread and sending awy an LdapConn
or an EntryStream is prohibited by the type system.