Add Search adapter framework
Adapters are an inversion-of-control mechanism to enable various
transformations of Search operations on the fly. The intent is to
enable the development of user-supplied adapters for complex and
environment-specific operations such as referral chasing without
the need to modify the base library.

Two adapters are supplied with the crate: EntriesOnly, mimicking
the earlier behavior of Search, which collected all referrals in
the end result, and PagedResults, which used the eponymous control
to automatically fetch large result sets in multiple Search
operations.