Add Kerberos/GSSAPI support
The support is off by default and requires the "gssapi" crate
feature. The user-visible interface is the Ldap::sasl_gssapi_bind()
method and its synchronous equivalent in LdapConn.

The bind operation automatically chooses the requested security
layer depending on the type of the connection: TLS connections get
no security layer, while clear connections get confidentiality
protection.

This version only knows to use the default Kerberos credentials
and requires explicitly supplying the FQDN of the LDAP server to
construct the SPN.