Use RwLock instead of atomics for SASL parameters
In addition to the SASL wrapping flag, the max GSSAPI wrapping buffer
size must be shared between Ldap and LdapCodec, which is easier to do
with a single struct behind a RwLock as opposed to multiple atomic
values. The path where the values are used is not performance
sensitive (right before network traffic).