Improve LDAP primer
A minor point was added about LDIF, just to avoid readers thinking servers held their
databases in that form.

I tried to make it clear that dn is not an attribute. This is a really common
misconception and it could surprise callers looking for the DN in the returned attrs.

In the filter section I added a note about other filter types. I also clarified the
slightly confusing examples searching for "john.doe" explaining which attributes matched.

I thought it was useful to note that LDAP is asynchronous, so clients can issue multiple
operations at once.

I added a small section on controls and extended operations.

Lots of other small little improvements.