Silence nightly warning
Probably due to rust-lang/rust#118297, our SaslCreds tuple struct
now generates a warning. It can be silenced by making the struct
public, but it really doesn't need to be, so dead code will be
allowd at this point.

The warning will presumably end up in stable Rust in a couple of
releases.