Fix Windows build with GSSAPI
The forcing of libgssapi when the "gssapi" feature is selected,
necessary for the selection of the appropriate libgssapi version
to fix the runtime panic (see commit 512401), broke the Windows
build, since it must use the platform SSPI libraries.

Since there are no platform-specific features, conditional optional
crate selection is used as a hack by superfluous feature setting
for libgssapi to include it in the build (with the fixed version)
only if cross-krb5 transitively selects it on Unix systems.